// src/Modules/changePassword/__generated__/changePassword.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: changePassword
// ====================================================

export interface changePassword {
  changePassword: boolean | null;
}

export interface changePasswordVariables {
  id: string;
  password: string;
}